backend/src/api/get_gpv_data_api/requirements.txt
functions-framework==3.*
pytz
google-cloud-datastore
google-cloud-storage
zstandard
numpy
google-auth
requests